refactor(esci2): lift Frida replay helpers into shared test-support (issue #27)
